+++ src/circuit_weaver/__init__.py
@@ -36,4 +36,4 @@ __all__ = [
     "update_board_placements",
 ]
 
-__version__ = "0.28.3"
+__version__ = "0.29.0"